+++ example/lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:quickly/quickly.dart';
 
 void main() {
   runApp(const MyApp());
@@ -10,16 +9,12 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return const MaterialApp(
       home: Scaffold(
         body: SafeArea(
-          child: Text('Welcome to Quickly'.toTitleCase)
-              .h1
-              .red500
-              .bold
-              .underline
-              .italic
-              .center,
+          child: Center(
+            child: Text('Quickly'),
+          ),
         ),
       ),
     );