+++ lib/quickly.dart
@@ -5,3 +5,5 @@ export 'src/dart/extensions.dart';
 export 'src/flutter/utilities.dart';
 
 export 'src/flutter/extensions.dart';
+
+export 'src/flutter/responsive.dart';