+++ lib/src/dart/num_extension.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'extensions.dart';
 
 ///Num extension to extend List functionality
@@ -13,18 +12,6 @@ extension NumExtension on num {
       3: '${this}rd',
     }.match(this % 10, '${this}th');
   }
-
-  ///Creates a fixed sized square box.
-  SizedBox sq(Widget child) => _sizedBox(child, h: toDouble(), w: toDouble());
-
-  ///Creates a fixed sized height box.
-  SizedBox h(Widget child) => _sizedBox(child, h: toDouble());
-
-  ///Creates a fixed sized width box.
-  SizedBox w(Widget child) => _sizedBox(child, w: toDouble());
-
-  SizedBox _sizedBox(Widget child, {double? h, double? w}) =>
-      SizedBox(child: child, height: h, width: w);
 
   ///Returns Duration of this in seconds.
   Duration get seconds => Duration(seconds: this as int);