+++ lib/src/flutter/extensions/num_extension.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 
-///Num extension to extend num functionality
+///Num extension on Widget
 extension NumExt on num {
   /// Creates a fixed square sized box.
-  SizedBox sq([Widget? child]) =>
+  SizedBox sqBox([Widget? child]) =>
       SizedBox(height: toDouble(), width: toDouble(), child: child);
 
   /// Create Fixed size width box
-  SizedBox w([Widget? child]) => SizedBox(width: toDouble(), child: child);
+  SizedBox wBox([Widget? child]) => SizedBox(width: toDouble(), child: child);
 
   /// Create Fixed size height box
-  SizedBox h([Widget? child]) => SizedBox(height: toDouble(), child: child);
+  SizedBox hBox([Widget? child]) => SizedBox(height: toDouble(), child: child);
 }