+++ lib/src/flutter/utilities.dart
@@ -1,7 +1,5 @@
 export 'utilities/padding.dart';
 
-export 'utilities/border_radius.dart';
+export 'utilities/radius.dart';
 
 export 'utilities/colors.dart';
-
-export 'utilities/responsive.dart';